+++ src/jerimum/Relogio.scala
@@ -4,6 +4,6 @@ import java.util.Date
 
 object Relogio {
   private[this] val inicio = new Date(1, 1, 2000).getTime;
-  def agora: Float = new Date().getTime
+  def agora: Double = new Date().getTime
   def milisegundos: Int = (System.currentTimeMillis() - inicio).toInt
 }
